权限限定 2020.4.14

+++ qyjsw/user/src/main/java/com/htqyjsw1/entity/TDept.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public class TDept  implements Serializable {
 
-    private Integer deptId;
+    private Long deptId;
 
     private String deptName;
 
@@ -17,11 +17,11 @@ public class TDept  implements Serializable {
     private String  deptTel;
 
 
-    public Integer getDeptId() {
+    public Long getDeptId() {
         return deptId;
     }
 
-    public void setDeptId(Integer deptId) {
+    public void setDeptId(Long deptId) {
         this.deptId = deptId;
     }
