+++ qyjsw/user/src/main/java/com/htqyjsw1/repository/UserRepository.java
@@ -74,7 +74,7 @@ public interface UserRepository {
      * @param deptId
      * @return list
      */
-    public List<TUser> queryUserByDeptId(Integer deptId) throws Exception;
+    public List<TUser> queryUserByDeptId(Long deptId) throws Exception;
 
     public List<TUser> queryUserByRoleId(Long roleId);
